+++ src/ggrc-client/js/plugins/utils/models-utils.js
@@ -5,7 +5,6 @@
 
 import {notifier} from './notifiers-utils';
 import RefreshQueue from '../../models/refresh_queue';
-import Stub from '../../models/stub';
 
 const relatedAssessmentsTypes = Object.freeze(['Control', 'Objective']);
 
@@ -80,145 +79,6 @@ const hasRelatedAssessments = (type) => {
   return _.includes(relatedAssessmentsTypes, type);
 };
 
-const handlePendingJoins = (obj) => {
-  const hasPendingJoins = _.get(obj, '_pending_joins.length') > 0;
-
-  if (!hasPendingJoins) {
-    return can.Deferred().resolve(obj);
-  }
-
-  return _resolveDeferredBindings(obj);
-};
-
-const resolveDeferredBindings = (obj) => {
-  const hasPendingJoins = _.get(obj, '_pending_joins.length') > 0;
-
-  if (!hasPendingJoins) {
-    return can.Deferred().resolve(obj);
-  }
-
-  return _resolveDeferredBindings(obj)
-    .then(() => obj.refresh());
-};
-
-function _resolveDeferredBindings(obj) {
-  const pendingJoins = obj._pending_joins.slice(0); // refresh of bindings later will muck up the pending joins on the object
-  const uniqueThrough = _(pendingJoins).chain()
-    .map((pj) => pj.through)
-    .uniq()
-    .value();
-
-  const refreshDfds = [];
-  can.each(uniqueThrough, (binding) => {
-    refreshDfds.push(obj.get_binding(binding).refresh_stubs());
-  });
-
-  return $.when(...refreshDfds)
-    .then(() => {
-      const dfds = [];
-      can.each(obj._pending_joins, (pj) => {
-        switch (pj.how) {
-          case 'add':
-            dfds.push(..._addHandler(obj, pj));
-            break;
-          case 'update':
-            dfds.push(..._updateHandler(obj, pj));
-            break;
-          case 'remove':
-            dfds.push(..._removeHandler(obj, pj));
-            break;
-          default: {
-            // nothing
-          }
-        }
-      });
-
-      const dfdsApply = $.when(...dfds);
-      obj.attr('_pending_joins', []);
-      obj.attr('_pending_joins_dfd', dfdsApply);
-
-      return dfdsApply.then(() => {
-        obj.dispatch('resolvePendingBindings');
-        return obj;
-      });
-    });
-}
-
-function _addHandler(obj, pj) {
-  const binding = obj.get_binding(pj.through);
-  const dfds = [];
-  // Don't re-add -- if the object is already mapped (could be direct or through a proxy)
-  // move on to the next one
-  if (_.includes(_.map(binding.list, 'instance'), pj.what) ||
-      (binding.loader.option_attr &&
-      _.includes(_.map(binding.list, function (joinObj) {
-        return joinObj.instance[binding.loader.option_attr];
-      }), pj.what))) {
-    return dfds;
-  }
-  const model = (CMS.Models[binding.loader.model_name] ||
-    GGRC.Models[binding.loader.model_name]);
-  const inst = pj.what instanceof model
-    ? pj.what
-    : new model({
-      context: obj.context,
-    });
-  const pjDfd = pj.what !== inst && pj.what.isNew()
-    ? pj.what.save() : null;
-  const dfd = $.when(pjDfd)
-    .then(function () {
-      if (binding.loader.object_attr) {
-        inst.attr(binding.loader.object_attr, new Stub(obj));
-      }
-      if (binding.loader.option_attr) {
-        inst.attr(binding.loader.option_attr, new Stub(pj.what));
-      }
-      if (pj.extra) {
-        inst.attr(pj.extra);
-      }
-      return inst.save();
-    });
-  dfds.push(dfd);
-  return dfds;
-}
-
-function _updateHandler(obj, pj) {
-  const dfds = [];
-  let binding = obj.get_binding(pj.through);
-  binding.list.forEach(function (boundObj) {
-    let blOptionAttr = binding.loader.option_attr;
-    if (boundObj.instance === pj.what ||
-        boundObj.instance[blOptionAttr] === pj.what) {
-      boundObj.get_mappings().forEach(function (mapping) {
-        dfds.push(mapping.refresh().then(function () {
-          if (pj.extra) {
-            mapping.attr(pj.extra);
-          }
-          return mapping.save();
-        }));
-      });
-    }
-  });
-  return dfds;
-}
-
-function _removeHandler(obj, pj) {
-  const binding = obj.get_binding(pj.through);
-  const dfds = [];
-  can.map(binding.list, function (boundObj) {
-    let blOptionAttr = binding.loader.option_attr;
-    if (boundObj.instance === pj.what ||
-        boundObj.instance[blOptionAttr] === pj.what) {
-      can.each(boundObj.get_mappings(), function (mapping) {
-        dfds.push(mapping.refresh().then(function () {
-          mapping.destroy();
-        }));
-      });
-    }
-  });
-  return dfds;
-}
-
 function _getObjectTypeDecisionTree() { // eslint-disable-line
   let tree = {};
   let extensions = GGRC.extensions || [];
@@ -240,8 +100,6 @@ export {
   getModelInstance,
   hasRelatedAssessments,
   relatedAssessmentsTypes,
-  resolveDeferredBindings,
-  handlePendingJoins,
   makeModelInstance,
   inferObjectType,
 };
